Moved ie::Parameter to new API
