+++ inference-engine/src/inference_engine/include/openvino/runtime/variable_state.hpp
@@ -5,7 +5,7 @@
 /**
  * @brief A header file that provides VariableState
  *
- * @file variable_state.hpp
+ * @file openvino/runtime/variable_state.hpp
  */
 
 #pragma once
@@ -16,7 +16,7 @@
 #include <memory>
 #include <string>
 
-#include "common.hpp"
+#include "openvino/runtime/common.hpp"
 
 namespace InferenceEngine {
 class IVariableStateInternal;
